Add files via upload
this code is necessary to use the ROB-14451 motor drivers to use all of our Peristaltic pumps, our DC motor, and our linear actuator. The runtime for all pumps/motors will have to be determined through testing and measuring the liquid distributions/distance traveled. (We will likely want to run the DC motor at will and determine the shutoff time based off when we have dirt in the tubes -- possibly connect the runtime to a button somehow). All pin definitions are temporary values used during testing on an Arduino Uno and must be updated to fit the pinout of the Teensy 4.1 board.