Added Sensors and modified vaccum/carousel (#6)
Sensor code was updated (temp and humidity merged). The file directory was also changed so that everything is now under science. Changes were made to accommodate changes to hardware (dirt carousel limit switch, vacuumLinear limit switch). The servo for the vacuum was also added. The dig sequence has NOT been completely updated yet. Anything related to the auger was also renamed to be related to the vacuum.